% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdca.R
\name{assoc_strength}
\alias{assoc_strength}
\title{Association strength}
\usage{
assoc_strength(df, assocstr_digit = assocstr_digit)
}
\arguments{
\item{df}{A nested/list data frame containing input data for transforming the p-value of the Binomial Test into association strength value}

\item{assocstr_digit}{Integer for the floating points/digits of the \emph{Association Strength}. The default is \code{3L}, which is passed-on from \code{\link{mdca}}.}
}
\value{
A double numeric vector of association strength
}
\description{
A utility function to derive association strength value in a row-wise, tidy fashion.
    It is derived via log-transformation to the base of 10L of the Binomial Test p-value.
    The function is called internally by \code{\link{mdca}}.
}
