% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdca.R
\name{binomial_test}
\alias{binomial_test}
\title{Binomial test}
\usage{
binomial_test(df)
}
\arguments{
\item{df}{A nested/list data frame containing input data for the Binomial Test}
}
\value{
A double numeric vector of p-value from the Binomial Test
}
\description{
A utility function to perform one-tailed Binomial Test in a row-wise, tidy fashion.
    It is called internally by \code{\link{mdca}}.
}
