% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colloc_input_data}
\alias{colloc_input_data}
\title{Distinctive collocates data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27989 rows and 2 columns.}
\usage{
colloc_input_data
}
\description{
The input data for collocational analysis (4-window span collocates of the synonyms).
}
\keyword{datasets}
