% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_cm}
\alias{df_cm}
\title{Interrater data for conceptual metaphors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 170 rows and 7 columns.
}
\usage{
df_cm
}
\description{
The input data for interrater agreement trial for the conceptual metaphors.
The data include classifications in the pre-discussion and post discussion rounds.
}
\keyword{datasets}
