% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_cxn_pattern}
\alias{df_cxn_pattern}
\title{Interrater data for the constructional patterns}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 3 columns.
}
\usage{
df_cxn_pattern
}
\description{
The input data for interrater agreement trial for determining the constructional patterns of the happiness words in their concordance lines.
The data include classifications in the pre-discussion round only since we already arrived at a nearly perfect agreement before the discussion.
}
\keyword{datasets}
