% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_meta_use_1st}
\alias{df_meta_use_1st}
\title{Interrater data for metaphorical usages of the pattern (1st round)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 262 rows and 5 columns.
}
\usage{
df_meta_use_1st
}
\description{
The input data for interrater agreement trial for determining whether the constructional patterns of the happiness words are metaphorical or not.
The data include classifications in the pre-discussion round.
}
\keyword{datasets}
