% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_colloc_gloss}
\alias{dist_colloc_gloss}
\title{Gloss for the distinctive collocates}
\format{An object of class \code{character} of length 1.}
\usage{
dist_colloc_gloss
}
\description{
A tibble/data frame containing gloss for the distinctive collocates
}
\keyword{datasets}
