% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract.happy_result_list}
\alias{extract.happy_result_list}
\title{Extract items from happy result lists}
\usage{
\method{extract}{happy_result_list}(happy_result_list, item, ...)
}
\arguments{
\item{happy_result_list}{A `happy_result_list` object.}

\item{item}{Item to extract. One of: summary, extended.}
}
\description{
Extract items across multiple happy result objects and combine into a single `data.frame`.
}

