% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{extract}
\alias{extract}
\title{Extract tables from hap.py result lists}
\usage{
extract(happy_result_list, item = c("summary", "extended"))
}
\arguments{
\item{happy_result_list}{A \code{happy_result_list} object, created
by combining \code{happy_result}s together with \code{c}}

\item{table}{Table of data to extract from each result}
}
\description{
Extract tables from multiple hap.py result objects and combine
into a single \code{data.frame}. Source information from each
result is added as an additional column (\code{from}).
}
