% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cic_meta_lu.R
\name{get_cic_meta_lu}
\alias{get_cic_meta_lu}
\title{Retrieve the conventionalised Lexical Units of a metaphor}
\usage{
get_cic_meta_lu(metaphor = NULL, df_ttr_out = NULL)
}
\arguments{
\item{metaphor}{regular expressions for the relevant metaphor.}

\item{df_ttr_out}{the tibble as the output of \code{\link{ttr}}.}
}
\value{
A tibble data frame
}
\description{
The function helps filter out LUs of a metaphor that can be considered as conventional linguistic instances (cf. Vergara Wilson, 2014, for the way to determine CIC, which stands for \emph{"conventionalised instances of constructions"}).
}
\examples{
ttr_metaphor <- ttr(df = phd_data_metaphor,
                    metaphor_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)

get_cic_meta_lu("possessable", ttr_metaphor)

}
\references{
Vergara Wilson, D. (2014). \emph{Categorization and Constructional Change in Spanish Expressions of "becoming"}. Leiden: Brill.
}
