% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frames.R
\name{get_frames}
\alias{get_frames}
\title{Retrieve the descriptive distributional statistics of source frames}
\usage{
get_frames(metaphor = NULL, word = NULL, df = NULL,
  frame_var = "source_frames", metaphor_var = "metaphors",
  synonym_var = "synonyms", lexunit_var = "lu")
}
\arguments{
\item{metaphor}{character strings of regular expressions for the metaphor.}

\item{word}{character strings of regular expressions for the happiness synonyms.}

\item{df}{the data frame (\code{phd_data_metaphor}).}

\item{frame_var}{character string for the column name of the source frame variable (i.e., \code{"source_frames"}).}

\item{metaphor_var}{character string for the column name of the metaphor variable (i.e., \code{"metaphors"}).}

\item{synonym_var}{character string for the column name of the synonyms variable (i.e., \code{"synonyms"}).}

\item{lexunit_var}{character string for the column name of the lexical unit variable (i.e., \code{"lu"}).}
}
\value{
A tibble/data frame of the descriptive statistics of the frames,
    namely the raw and relative (i.e. percentages of the) token and type frequencies.
    If the \code{word} argument is specified,
    both the \code{source_frames} and the \code{synonyms} variable are grouped
    for calculating the token and type frequencies, while percentages of these are summarised per word.
}
\description{
A utility function to get the descriptive data for the source frames of a metaphor (for the aggregated synonyms data or for a given synonym, if specified).
   This function is used in Rajeg's (2018, Chapter 5 & 6).
}
\examples{

# for aggregated data across all synonyms
get_frames(metaphor = "desired goal",
           df = phd_data_metaphor)

# for a synonym
get_frames(metaphor = "desired goal",
           word = "^kesenangan",
           df = phd_data_metaphor)

# for synonyms with the same root (e.g. "senang" and "kesenangan")
get_frames(metaphor = "desired goal",
           word = "senang",
           df = phd_data_metaphor)

# for synonyms in the nominalised forms
get_frames(metaphor = "desired goal",
           word = "^(kesenangan|kegembiraan)",
           df = phd_data_metaphor)

}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for happiness} (PhD Thesis). Monash University. Melbourne, Australia.
}
