% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lexically_diverse_metaphors.R
\name{get_lexically_diverse_metaphors}
\alias{get_lexically_diverse_metaphors}
\title{Retrieve the top-10 lexically diverse metaphors}
\usage{
get_lexically_diverse_metaphors(df_ttr_out = NULL, min_token = 3,
  top_n_limit = 10)
}
\arguments{
\item{df_ttr_out}{the tibble as the output of \code{\link{ttr}}.}

\item{min_token}{minimum token frequency of the metaphors.}

\item{top_n_limit}{limit for presenting the top-n most lexically diverse metaphors.}
}
\value{
A tibble data frame.
}
\description{
This function calls the script to output the top-10 metaphors that have high index of lexical creativity in their linguistic manifestation.
    The lexical creativity is derived from the \code{type-per-token ratio (TTR)} measure (Rajeg, 2018, Chapter 6).
}
\examples{
# get the ttr table
ttr_metaphor <- ttr(phd_data_metaphor,
                    schema_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)

# get the lexically diverse metaphors
get_lexically_diverse_metaphors(df_ttr_out = ttr_metaphor,
                                min_token = 3,
                                top_n_limit = 10)
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS} (PhD Thesis). Monash University. Melbourne, Australia.
}
