% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lu_frame_table.R
\name{get_lu_frame_table}
\alias{get_lu_frame_table}
\title{Retrieve the (top-n most frequent) Lexical Units evoking the source frames of a metaphor}
\usage{
get_lu_frame_table(
  metaphor = NULL,
  word = NULL,
  print_all = FALSE,
  top_n = FALSE,
  limit = 10,
  df = NULL,
  metaphor_var = "metaphors",
  frame_var = "source_frames",
  synonym_var = "synonyms",
  lexunit_var = "lu",
  lexunit_gloss_var = "lu_gloss"
)
}
\arguments{
\item{metaphor}{regular expressions for the relevant metaphor.}

\item{word}{character strings of regular expressions for the HAPPINESS synonyms.}

\item{print_all}{logical; print all output}

\item{top_n}{logical; limit to the top-n LUs}

\item{limit}{integer; maximum rows to be printed from the top-n LUs}

\item{df}{the data frame (\code{phd_data_metaphor}).}

\item{metaphor_var}{character string for the column name of the metaphor variable (i.e., \code{"metaphors"}).}

\item{frame_var}{character string for the column name of the source frame variable (i.e., \code{"source_frames"}).}

\item{synonym_var}{character string for the column name of the synonyms variable (i.e., \code{"synonyms"}).}

\item{lexunit_var}{character string for the column name of the lexical unit variable (i.e., \code{"lu"}).}

\item{lexunit_gloss_var}{character string for the column name of the gloss of the lexical units (i.e., \code{"lu_gloss"}).}
}
\value{
A tibble/data frame of the descriptive statistics of the LUs and the frames, including the gloss of the LUs.
}
\description{
The function is designed to easily retrieve a tibble data frame of the top-n most frequent frame-evoking Lexical Units (LU) of the prominent metaphors discussed in Rajeg (2019, Chapter 5 and 6).
}
\examples{
# for the aggregated data
get_lu_frame_table(metaphor = "possessable",
                   top_n = TRUE,
                   limit = 20,
                   df = phd_data_metaphor)

# for a specific synonym
get_lu_frame_table(metaphor = "possessable",
                   word = "^bahagia$",
                   top_n = TRUE,
                   limit = 20,
                   df = phd_data_metaphor)
}
\references{
Rajeg, G. P. W. (2019). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS} (PhD Thesis). Monash University. Melbourne, Australia. \url{https://doi.org/10.26180/5cac231a97fb1}.
}
