% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lu_submappings_table.R
\name{get_lu_submappings_table}
\alias{get_lu_submappings_table}
\title{Retrieve the frequency of Lexical units and the evoked submappings}
\usage{
get_lu_submappings_table(metaphor = NULL, word = NULL,
  print_all = FALSE, top_n = FALSE, limit = 10, df = NULL,
  metaphor_var = "metaphors", submet_var = "submappings",
  synonym_var = "synonyms", lexunit_var = "lu",
  lexunit_gloss_var = "lu_gloss")
}
\arguments{
\item{metaphor}{regular expressions for the relevant metaphor.}

\item{word}{character strings of regular expressions for the HAPPINESS synonyms.}

\item{print_all}{logical; print all output}

\item{top_n}{logical; limit to the top-n LUs}

\item{limit}{integer; maximum rows to be printed from the top-n LUs}

\item{df}{the data frame (\code{phd_data_metaphor}).}

\item{metaphor_var}{character string for the column name of the metaphor variable (i.e., \code{"metaphors"}).}

\item{submet_var}{character string for the column name of the submapping variable (i.e., \code{"submappings"}).}

\item{synonym_var}{character string for the column name of the synonyms variable (i.e., \code{"synonyms"}).}

\item{lexunit_var}{character string for the column name of the lexical unit variable (i.e., \code{"lu"}).}

\item{lexunit_gloss_var}{character string for the column name of the gloss of the lexical units (i.e., \code{"lu_gloss"}).}
}
\value{
A tibble containing the submappings, LUs and the glosses,
    frequency of each LUs of the metaphor,
    percentage of each LUs in the metaphor,
    and percentage of each LUs by the evoked submappings.
}
\description{
The function generates a table of frequency and percentages of Lexical units evoking certain submappings of a metaphor.
}
\examples{
# for all synonyms
get_lu_submappings_table(metaphor = "liquid in a container",
                         df = phd_data_metaphor)

# for \\emph{kegembiraan} 'joy'
get_lu_submappings_table(metaphor = "liquid in a container",
                         word = "kegembiraan",
                         df = phd_data_metaphor)
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS}. PhD Thesis. Monash University. Melbourne, Australia.
}
