% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lu_table.R
\name{get_lu_table}
\alias{get_lu_table}
\title{Retrieve the (top-n most frequent) Lexical Units of a metaphor}
\usage{
get_lu_table(metaphor = "regular expressions", top_n_only = TRUE,
  top_n_limit = 10, df = NULL, submapping_perc = FALSE,
  incl_submappings = FALSE)
}
\arguments{
\item{metaphor}{regular expressions for the relevant metaphor.}

\item{top_n_only}{logical; the default is \code{TRUE} with the number specified in \code{top_n_limit} argument. If \code{FALSE}, it prints all the LU for the metaphor.}

\item{top_n_limit}{integer; how many most frequent LU should be printed? The default is \code{10}.}

\item{df}{the data frame for the thesis (\code{phd_data_metaphor}). This argument is not put first because the main attention when using the function is on varying the metaphor (hence, coming first) rather than on varying the database.}

\item{submapping_perc}{logical; wheter to include the percentages of the LU by submappings. If \code{TRUE}, the function only print out the percentages of the LU by submappings.
To include the submapping types, specify the \code{incl_submappings} argument to \code{TRUE} as well (the default is \code{FALSE}).}

\item{incl_submappings}{logical; whether to include the submappings inferred from the LU (\code{TRUE}) or not (\code{FALSE} -- the default).}
}
\value{
A tbl_df.
}
\description{
The function is designed to easily retrieve a tibble data frame of the top-n most frequent Lexical Units (LU) of the prominent metaphors discussed in Chapter 5 and 6.
}
\examples{
get_lu_table(metaphor = "destination", df = phd_data_metaphor)
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS}. PhD Thesis. Monash University. Melbourne, Australia.
}
