% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_freq_profile.R
\name{get_meta_freq_profiles}
\alias{get_meta_freq_profiles}
\title{Select a (set of) frequency profile columns for a metaphor}
\usage{
get_meta_freq_profiles(metaphor = NULL, ..., metaphor_var = "metaphors",
  df = NULL)
}
\arguments{
\item{metaphor}{character string of regular expressions for the metaphors.}

\item{...}{any bare variable/column name of the frequency profiles from \code{ttr} function.}

\item{metaphor_var}{character string for the column name of the metaphor variable (i.e., \code{"metaphors"}).}

\item{df}{data frame storing the results of \code{ttr}.}
}
\value{
A tibble data frame
}
\description{
Select a (set of) frequency profile columns for a metaphor
}
\examples{
# perform ttr
ttr_metaphor <- ttr(phd_data_metaphor,
                    metaphor_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)

# retrieve the profile
get_meta_freq_profiles("food", # regex for the metaphor
                       token, type_lu, type_per_token_lu, # inputs for the dot-dot-dot (...)
                       metaphor_var = "metaphors",
                       df = ttr_metaphor)
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for happiness} (PhD Thesis). Monash University. Melbourne, Australia.
}
