% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submappings.R
\name{get_submappings}
\alias{get_submappings}
\title{Retrieve the submappings of the metaphors overall or by synonyms/words}
\usage{
get_submappings(metaphor = NULL, word = NULL, df = NULL,
  metaphor_var = "metaphors", submet_var = "submappings",
  synonym_var = "synonyms", lexunit_var = "lu")
}
\arguments{
\item{metaphor}{character strings of regular expressions for the metaphor.}

\item{word}{character strings of regular expressions for the happiness synonyms.}

\item{df}{the data frame (\code{phd_data_metaphor}).}

\item{metaphor_var}{character string for the column name of the metaphor variable (i.e., \code{"metaphors"}).}

\item{submet_var}{character string for the column name of the submapping variable (i.e., \code{"submappings"}).}

\item{synonym_var}{character string for the column name of the synonyms variable (i.e., \code{"synonyms"}).}

\item{lexunit_var}{character string for the column name of the lexical unit variable (i.e., \code{"lu"}).}
}
\value{
A tibble/data frame of the descriptive statistics of the submappings, namely the raw and relative (i.e. percentages of the) token and type frequencies.
}
\description{
A utility function to get the descriptive data for the submappings of a metaphor (for the aggregated synonyms data or for a given synonym, if specified).
   This function is used in Rajeg's (2018, Chapter 5 & 6).
}
\examples{
submet <- get_submappings(metaphor = "possessable object$", df = phd_data_metaphor)

}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS} (PhD Thesis). Monash University. Melbourne, Australia.
}
