% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{this is the title}
\usage{
hello(string = "Heike")
}
\arguments{
\item{string}{character value as greeting, defaults to "Heike"}
}
\description{
this is a description of what the function does
\eqn{\frac{1}{n}\sum_{i=1}^n x_i}
}
\author{
ME
}
