% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdca_attr_repel.R
\name{mdca_repel}
\alias{mdca_repel}
\title{Repelled items from MDCA}
\usage{
mdca_repel(df = NULL, filter_by = "cxn", cxn_var = "synonyms",
  cxn_type = NULL, coll_var = NULL, coll_type = NULL, min_collstr = -2)
}
\arguments{
\item{df}{output data frame from \code{link{mdca}}.}

\item{filter_by}{character string indicating whether the output of \code{link{mdca}} is filtered according to the \emph{construction} variable (i.e. \code{"cxn"}) or the \emph{collexemes/collocates} variable (i.e. \code{"colloc"}).
If it is left \code{NULL} (the default), the data is filtered according to the value specified in the \code{min_collstr} argument.}

\item{cxn_var}{character string for the column name of the construction variable (i.e., \code{"synonyms"}).}

\item{cxn_type}{character strings of regular expressions for the HAPPINESS synonyms.}

\item{coll_var}{character string for the column name of the metaphor variable (i.e., \code{"metaphors"}).}

\item{coll_type}{character strings of regular expressions for the conceptual metaphors.}

\item{min_collstr}{numeric threshold for the minimum collostruction strength. By default it is set to \code{-2}.}
}
\value{
A tibble/data frame
}
\description{
A function to retrieve list of \emph{repelled} items for a construction from the results of \code{\link{mdca}}.
}
\examples{
# for distinctive metaphors
mdca_res <- mdca(df = phd_data_metaphor,
                 cxn_var = "synonyms",
                 coll_var = "metaphors",
                 correct_holm = TRUE,
                 concise_output = TRUE,
                 already_count_table = FALSE,
                 collstr_digits = 3)
# retrieve the repelled metaphors for "kebahagiaan".
mdca_repel(df = mdca_res,
          filter_by = "cxn",
          cxn_var = "synonyms",
          cxn_type = "kebahagiaan",
          min_collstr = -2)

}
