% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{mymean}
\alias{mymean}
\title{Calculate the average of a numeric vector}
\usage{
mymean(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{boolean value}
}
\value{
double value
}
\description{
Calculate the average of a numeric vector
}
