% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Turn numbers into words}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{Numeric integer to be converted into text/strings}
}
\value{
Character strings representing numbers
}
\description{
A function for turning numeric input into its corresponding character string.
}
\examples{
numbers2words(2L)
}
\references{
\itemize{
        \item http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html
        \item https://gist.github.com/psychemedia/150cb9901529da58124a
    }
}
