% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_body_part.R
\name{plot_body_part}
\alias{plot_body_part}
\title{Plot most frequent body-part terms}
\usage{
plot_body_part(df = NULL)
}
\arguments{
\item{df}{the data frame for the thesis (\code{phd_data_metaphor}).}
}
\value{
Barplot based on \code{ggplot2} package
}
\description{
A function to call the script to generate the Body-part barplot in Rajeg (2019, p. 163, Figure 5-1).
}
\examples{
plot_body_part(df = phd_data_metaphor)

}
\references{
Rajeg, G. P. W. (2019). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS} (PhD Thesis). Monash University. Melbourne, Australia. \url{https://doi.org/10.26180/5cac231a97fb1}.
}
