% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cxn_interrater.R
\name{plot_cxn_interrater}
\alias{plot_cxn_interrater}
\title{Top constructional patterns}
\usage{
plot_cxn_interrater(df = NULL)
}
\arguments{
\item{df}{data frame for the plot included in the package, that is \code{top_cxn_data}.}
}
\value{
A ggplot image
}
\description{
A function to generate Figure 3.1 in Rajeg (2019).
    The figure shows the proportion of the most frequent
    constructional patterns agreed during the interrater agreement trial.
}
\examples{
plot_cxn_interrater(df = top_cxn_data)


}
\references{
Rajeg, G. P. W. (2019). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS}. PhD Thesis. Monash University. Melbourne, Australia. \url{https://doi.org/10.26180/5cac231a97fb1}.
}
