% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_happy}
\alias{read_happy}
\title{Load a hap.py results directory}
\usage{
read_happy(happy_prefix, lazy = TRUE)
}
\arguments{
\item{happy_prefix}{hap.py output prefix (and path)}

\item{lazy}{store lesser-used output as
unevaluated promises rather than reading
everything at once}
}
\value{
A list structure containing hap.py output
}
\description{
Read a directory of hap.py output files into
an R data structure.
}
\examples{

\dontrun{
# run hap.py from commandline:
#  hap.py truth.vcf query.vcf -o /output/path/prefix

# load result into R
happy <- read_happy('/output/path/prefix')
names(happy)
}

}
\seealso{
read_sompy
}
