% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_samplesheet}
\alias{read_samplesheet}
\title{Load hap.py results in bulk from a samplesheet.csv}
\usage{
read_samplesheet(samplesheet_path, lazy = TRUE)
}
\arguments{
\item{samplesheet_path}{Path to samplesheet.csv. Required fields: `replicate_id`, `happy_prefix`.}

\item{lazy}{Do not load larger hap.py results until needed. Default: `TRUE`.}
}
\value{
A `happy_samplesheet` object, with the following fields:
\itemize{
  \item{`samplesheet`: the original samplesheet, stored as a `data.frame`}.
  \item{`results`: a `happy_result_list` that contains individual `happy_result` objects}.
}
}
\description{
Load multiple hap.py results into a `happy_samplesheet` object starting from a samplesheet.csv.
}
\examples{

\dontrun{
hap_samplesheet <- read_samplesheet(samplesheet_path = 'happyr_samplesheet.csv')
}

}
