% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_samplesheet_}
\alias{read_samplesheet_}
\title{Load hap.py results in bulk from a samplesheet dataframe}
\usage{
read_samplesheet_(samplesheet, lazy = TRUE)
}
\arguments{
\item{samplesheet}{A `happyR` samplesheet (`data.frame`). Required fields: `replicate_id`, `happy_prefix`.}

\item{lazy}{Do not load larger hap.py results until needed. Default: `TRUE`.}
}
\value{
A `happy_samplesheet` object, with the following fields:
\itemize{
  \item{`samplesheet`: the original samplesheet, stored as a `data.frame`}.
  \item{`results`: a `happy_result_list` that contains individual `happy_result` objects}.
}
}
\description{
Load multiple hap.py results into a `happy_samplesheet` object starting from a samplesheet dataframe.
}
\examples{

\dontrun{
samplesheet <- readr::read_csv("group_id,replicate_id,happy_prefix
PCR-Free,NA12878-I30,NA12878-I30_S1
PCR-Free,NA12878-I33,NA12878-I33_S1
Nano,NA12878-R1,NA12878-R1_S1
Nano,NA12878-R2,NA12878-R2_S1
")
hap_samplesheet <- read_samplesheet_(samplesheet = samplesheet_df)
}

}
