% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_sompy}
\alias{read_sompy}
\title{Load a som.py results directory}
\usage{
read_sompy(sompy_prefix, lazy = TRUE)
}
\arguments{
\item{sompy_prefix}{som.py output prefix (and path)}

\item{lazy}{store lesser-used output as
unevaluated promises rather than reading
everything at once}
}
\value{
A list structure containing som.py output
}
\description{
Read a directory of som.py (somatic hap.py) output
files into an R data structure.
}
\examples{

\dontrun{
# run som.py from commandline:
#  som.py truth.vcf query.vcf -o /output/path/prefix

# load result into R
sompy <- read_sompy('/output/path/prefix')
names(sompy)
}

}
\seealso{
read_happy
}

