% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum3.R
\name{sum3}
\alias{sum3}
\title{Add together two numbers.}
\usage{
sum3(x, y, z)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{z}{A number.}
}
\value{
The sum of \code{x} and \code{y} and \code{z}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1,2)
add(10, 1,3)
}
