% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_cxn_data}
\alias{top_cxn_data}
\title{Interrater data for the top constructional patterns}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.
}
\usage{
top_cxn_data
}
\description{
The input data for \emph{Figure 3-1} in Rajeg (2018).
The barplot shows the distribution of constructional patterns during interrater agreement trial whose frequency of occurrences are at least five (5).
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS}. PhD Thesis. Monash University. Melbourne, Australia.
}
\keyword{datasets}
