% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttr.R
\name{ttr}
\alias{ttr}
\title{Perform the frequency profile measures in Chapter 5 and Chapter 6}
\usage{
ttr(df = NULL, metaphor_var = "metaphors", lexunit_var = "lu",
  float_digits = 2)
}
\arguments{
\item{df}{the data frame for the database of the thesis (\code{phd_data_metaphor.rda}).}

\item{metaphor_var}{character string of the column name for the metaphor variable in the data frame (i.e., \code{"metaphors"}).}

\item{lexunit_var}{character string of the column name for the lexical unit variable in the data frame (i.e., \code{"lu"}).}

\item{float_digits}{integer indicating the retained floating points from the calculation. The default is \code{2}.}
}
\value{
A tibble data frame (\code{tbl_df}) sorted in decreasing order of the \code{token} frequency of the metaphors.
}
\description{
Generate a set of basic frequency profiles for the metaphors: (i) token frequency, (ii) type frequency, and (iii) type per token ratio (TTR).
}
\examples{
ttr_metaphor <- ttr(df = phd_data_metaphor,
                    metaphor_var = "metaphors",
                    lexunit_var = "lu",
                    float_digits = 2)
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for happiness}. PhD Thesis. Monash University. Melbourne, Australia.
}
