% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colloc_input_data}
\alias{colloc_input_data}
\title{Distinctive collocates data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27989 rows and 2 columns.
}
\usage{
colloc_input_data
}
\description{
The input data for distinctive collocates analysis in Rajeg (2019, Ch.7). The collocates data are words co-occurring within the span of 4 words to the right and left of the happiness near-synonyms. Following Gevaert (2007, p. 197), the collocates that are part of the studied synonyms are excluded because they may neutralise the distinctive features of the near-synonyms when the goal is to find semantic differences between these near-synonyms.
}
\details{
The data consists of the following two variables:
\describe{
\item{synonyms}{The ten near-synonyms under study. Note that only results of the nominalised, \emph{ke- -an} forms are presented in the thesis, but not the root forms.}
\item{collocates}{The window-span collocates of the synonyms}
}
}
\references{
\itemize{
\item{Rajeg, G. P. W. (2019). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for HAPPINESS} (PhD Thesis). Monash University. Melbourne, Australia. https://doi.org/10.26180/5cac231a97fb1}
\item{Gevaert, C. (2007). \emph{The history of ANGER: The lexical field of ANGER from Old to Early Modern English} (PhD thesis). Katholieke Universiteit Leuven, Leuven.
}
}
}
\keyword{datasets}
