% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_colloc_gloss}
\alias{dist_colloc_gloss}
\title{Gloss for the distinctive collocates}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 234 rows and 2 columns.}
\usage{
dist_colloc_gloss
}
\description{
A tibble data frame containing gloss for the distinctive collocates of the synonyms presented in Rajeg (2018, Ch.7).
}
\references{
Rajeg, G. P. W. (2018). \emph{Metaphorical profiles and near-synonyms: A corpus-based study of Indonesian words for happiness} (PhD Thesis). Monash University. Melbourne, Australia.
}
\keyword{datasets}
