% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{extract_results}
\alias{extract_results}
\title{Extract tables from hap.py result lists}
\usage{
extract_results(happy_result_list, table = c("summary", "extended",
  "pr.all", "pr.indel.pass", "pr.indel.sel", "pr.indel.all", "pr.snp.pass",
  "pr.snp.sel", "pr.snp.all"))
}
\arguments{
\item{happy_result_list}{A \code{happy_result_list} object, created
by combining \code{happy_result}s together with \code{c}}

\item{table}{Table of data to extract from each result.
\code{"summary"} or \code{"extended"} get top level tables
while the \code{pr} options get Precision-Recall tables.}
}
\value{
a \code{data.frame} of combined tables from list
}
\description{
Extract results tables from multiple hap.py result objects and combine
into a single \code{data.frame}. Source information from each
result is added as an additional column (\code{from}).
}
\examples{

\dontrun{
happy1 <- read_happy('/output/path/prefix')
happy2 <- read_happy('/different/path/prefix')
results_list <- c(happy1, happy2)

# get full extended metrics for all results as a data.frame
extended_df <- extract_results(results_list, table = "extended")

# get collapsed summary table of high-level metrics
summary_df <- extract_results(results_list, table = "summary")
unique(summary_df$from)
# [1] "/output/path/prefix"  "/different/path/prefix"
}

}
