% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phd_data_metaphor}
\alias{phd_data_metaphor}
\title{The metaphor data frame}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3638 rows and 10 columns.}
\usage{
phd_data_metaphor
}
\description{
The metaphor data for TTR analysis in Ch. 5 & Ch. 6, and for the MDCA analysis in Ch. 7.
}
\keyword{datasets}
