% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.happy_result}
\alias{print.happy_result}
\title{Print a hap.py results object}
\usage{
\method{print}{happy_result}(happy_result, ...)
}
\arguments{
\item{happy_result}{an object of class happy_result}

\item{...}{additional args passed down to \code{tibble::trunc_mat}}
}
\description{
Pretty-print the contents of a list-style
object containing hap.py result data
}
\examples{

\dontrun{
# not run example:
happy <- read_happy('/output/path/prefix')
print(happy)
}

# running example with package demo data
happy_input <- system.file("extdata", "happy_demo.summary.csv", package = "happyR")
happy_prefix <- sub(".summary.csv", "", happy_input)

# load happy result
hapdata <- read_happy(happy_prefix)
print(hapdata)

}
\seealso{
\code{\link[tibble]{trunc_mat}}
}
