% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_happy}
\alias{read_happy}
\title{Load a hap.py results directory}
\usage{
read_happy(directory, lazy = TRUE)
}
\arguments{
\item{directory}{path to the output directory}

\item{lazy}{store lesser-used output as
unevaluated promises rather than reading
everything}
}
\value{
A list structure containing hap.py output
}
\description{
Read a directory of hap.py output files into
a R data structure
}
\examples{

\dontrun{
happy <- read_happy('/path/to/output/dir')
names(happy)
}

}
