% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_samplesheet}
\alias{read_samplesheet}
\title{Load hap.py results from a samplesheet csv}
\usage{
read_samplesheet(samplesheet_path, lazy = TRUE)
}
\arguments{
\item{samplesheet_path}{Path to samplesheet.csv. Required fields: \code{replicate_id}, \code{happy_prefix}.}

\item{lazy}{Do not load larger hap.py results until needed. Default: \code{TRUE}.}
}
\value{
A \code{happy_samplesheet} object, with the following fields:
\itemize{
  \item{\code{samplesheet}: the original samplesheet, stored as a \code{data.frame}}.
  \item{\code{results}: a \code{happy_result_list} that contains individual \code{happy_result} objects}.
}
}
\description{
Load multiple hap.py results into a \code{happy_samplesheet} object starting from a samplesheet csv.
}
\examples{

\dontrun{
hap_samplesheet <- read_samplesheet(samplesheet_path = 'happyr_samplesheet.csv')
}

}
