% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaps.R
\name{scaps}
\alias{scaps}
\title{Small-capital}
\usage{
scaps(x = NULL)
}
\arguments{
\item{x}{A character strings to be turned into small capitals}
}
\value{
A character vector
}
\description{
A utility function to turn lower case character into small capitals
}
\examples{
scaps("happiness is a desired goal")
}
\references{
Rajeg, G. P. W. (2018). Scapr [R]. \url{https://doi.org/10.5281/zenodo.1327273}. \url{https://gederajeg.github.io/scapr/} (Original work published 2018)
}
